refactor: int형 double로 변경

+++ src/main/java/janggi/domain/piece/PieceType.java
@@ -15,19 +15,19 @@ import java.util.List;
 
 public enum PieceType {
 
-    GENERAL(new PalaceStrategy(), new EmptyCondition(), 0),
-    CHARIOT(new MultiStepStraightStrategy(), new EmptyCondition(), 13),
-    CANNON(new MultiStepStraightStrategy(), new OnePieceExistsCondition(), 7),
-    HORSE(new HorseStrategy(), new EmptyCondition(), 5),
-    ELEPHANT(new ElephantStrategy(), new EmptyCondition(), 3),
-    GUARD(new PalaceStrategy(), new EmptyCondition(), 3),
-    SOLDIER(new SoldierStrategy(), new EmptyCondition(), 2);
+    GENERAL(new PalaceStrategy(), new EmptyCondition(), 0.0),
+    CHARIOT(new MultiStepStraightStrategy(), new EmptyCondition(), 13.0),
+    CANNON(new MultiStepStraightStrategy(), new OnePieceExistsCondition(), 7.0),
+    HORSE(new HorseStrategy(), new EmptyCondition(), 5.0),
+    ELEPHANT(new ElephantStrategy(), new EmptyCondition(), 3.0),
+    GUARD(new PalaceStrategy(), new EmptyCondition(), 3.0),
+    SOLDIER(new SoldierStrategy(), new EmptyCondition(), 2.0);
 
     private final MoveStrategy moveStrategy;
     private final MoveCondition moveCondition;
     private final double score;
 
-    PieceType(MoveStrategy moveStrategy, MoveCondition moveCondition, int score) {
+    PieceType(MoveStrategy moveStrategy, MoveCondition moveCondition, double score) {
         this.moveStrategy = moveStrategy;
         this.moveCondition = moveCondition;
         this.score = score;
